share types
