+++ webui/react/src/components/LogViewer/LogViewer.tsx
@@ -14,11 +14,13 @@ import useResize from 'hooks/useResize';
 import { FetchArgs } from 'services/api-ts-sdk';
 import { readStream } from 'services/utils';
 import Message, { MessageType } from 'shared/components/message';
-import { Log, LogLevel, RecordKey } from 'types';
+import { Log, LogLevel } from 'types';
 import { clone } from 'utils/data';
 import { formatDatetime } from 'utils/datetime';
 import { copyToClipboard } from 'utils/dom';
 import { dateTimeStringSorter, numericSorter } from 'utils/sort';
+
+import { RecordKey } from '../../tmp-shared/types';
 
 import css from './LogViewer.module.scss';
 import LogViewerEntry, { DATETIME_FORMAT, ICON_WIDTH, MAX_DATETIME_LENGTH } from './LogViewerEntry';
