+++ webui/react/src/components/ResponsiveTable.test.tsx
@@ -5,9 +5,10 @@ import { FilterDropdownProps } from 'antd/es/table/interface';
 import React from 'react';
 
 import { getFullPaginationConfig, MINIMUM_PAGE_SIZE } from 'components/Table';
-import { Pagination, RecordKey, UnknownRecord } from 'types';
 import { alphaNumericSorter, numericSorter } from 'utils/sort';
 import { generateAlphaNumeric } from 'utils/string';
+
+import { Pagination, RecordKey, UnknownRecord } from '../tmp-shared/types';
 
 import ResponsiveTable from './ResponsiveTable';
 import TableFilterDropdown, {
