+++ webui/react/src/components/UPlot/UPlotScatter/UPlotScatter.utils.ts
@@ -1,8 +1,8 @@
 import uPlot from 'uplot';
 
-import { Range } from 'types';
 import { rgba2str, rgbaFromGradient, str2rgba } from 'utils/color';
 
+import { Range } from '../../../tmp-shared/types';
 import { UPlotAxisSplits, UPlotData } from '../types';
 
 export const X_INDEX = 0;
