+++ webui/react/src/hooks/useSettings.ts
@@ -2,10 +2,11 @@ import queryString from 'query-string';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import { useHistory, useLocation } from 'react-router-dom';
 
-import { Primitive, RecordKey } from 'types';
 import { clone, hasObjectKeys, isBoolean, isEqual, isNumber, isString } from 'utils/data';
 import handleError, { ErrorType } from 'utils/error';
 import { Storage } from 'utils/storage';
+
+import { Primitive, RecordKey } from '../tmp-shared/types';
 
 import usePrevious from './usePrevious';
 import useStorage from './useStorage';
