+++ webui/react/src/services/decoder.ts
@@ -5,6 +5,8 @@ import * as types from 'types';
 import { flattenObject, isNullOrUndefined, isNumber, isObject, isPrimitive } from 'utils/data';
 import { capitalize } from 'utils/string';
 
+import { Pagination, RawJson } from '../tmp-shared/types';
+
 import * as Sdk from './api-ts-sdk'; // API Bindings
 
 export const mapV1User = (data: Sdk.V1User): types.DetailedUser => {
@@ -22,7 +24,7 @@ export const mapV1UserList = (data: Sdk.V1GetUsersResponse): types.DetailedUser[
   return (data.users || []).map(user => mapV1User(user));
 };
 
-export const mapV1Pagination = (data: Sdk.V1Pagination): types.Pagination => {
+export const mapV1Pagination = (data: Sdk.V1Pagination): Pagination => {
   return {
     limit: data.limit ?? 0,
     offset: data.offset ?? 0,
@@ -408,9 +410,9 @@ export const mapV1ExperimentList = (data: Sdk.V1Experiment[]): types.ExperimentI
   return data.map(mapV1Experiment);
 };
 
-const filterNonScalarMetrics = (metrics: types.RawJson): types.RawJson | undefined => {
+const filterNonScalarMetrics = (metrics: RawJson): RawJson | undefined => {
   if (!isObject(metrics)) return undefined;
-  const scalarMetrics: types.RawJson = {};
+  const scalarMetrics: RawJson = {};
   for (const key in metrics) {
     if ([ 'Infinity', '-Infinity', 'NaN' ].includes(metrics[key])) {
       scalarMetrics[key] = Number(metrics[key]);
