+++ webui/react/src/types.ts
@@ -1,19 +1,10 @@
 import * as Api from 'services/api-ts-sdk';
 
+import { Primitive, RawJson, RecordKey } from './tmp-shared/types';
+
 interface WithPagination {
   pagination: Api.V1Pagination; // probably should use this or Pagination
 }
-
-export type RecordKey = string | number | symbol;
-export type UnknownRecord = Record<RecordKey, unknown>;
-export type Primitive = boolean | number | string;
-export type NullOrUndefined<T = undefined> = T | null | undefined;
-export type Point = { x: number; y: number };
-export type Range<T = Primitive> = [ T, T ];
-export type Eventually<T> = T | Promise<T>;
-
-/* eslint-disable-next-line @typescript-eslint/no-explicit-any */
-export type RawJson = Record<string, any>;
 
 export type PropsWithStoragePath<T> = T & { storagePath?: string };
 
@@ -139,11 +130,6 @@ export interface EndTimes {
 
 export interface StartEndTimes extends EndTimes {
   startTime: string;
-}
-
-export interface Pagination {
-  limit: number;
-  offset: number;
 }
 
 /* Command */
