+++ webui/react/src/utils/chart.ts
@@ -1,6 +1,7 @@
 import themes, { defaultThemeId } from 'themes';
-import { Primitive, Range } from 'types';
 import { primitiveSorter } from 'utils/sort';
+
+import { Primitive, Range } from '../tmp-shared/types';
 
 import { ColorScale } from './color';
 import { clone } from './data';
